requirements.txt
grpcio-tools ==1.59.0
grpcio ==1.59.0
protobuf == 4.24.4